test: Move isolated test to image tests
